test(grok-web): isolate stream boundary fixture
